machine/iot2000: remove settings to rebuild glibc from source
MEL used to ship a pre-built version of the glibc library. This was an issue on the
iot2000 as it has to be compiled with a special compiler flag (-Wa,-no-lock-prefix)
to work-around a hardware bug with the Intel Quark processor. To address this issue,
the meta-iot2000-bsp layer was instructing MEL to rebuild glibc from source (and use
the aforementioned compiler flag). Starting with MEL 2017.02 Update 2, glibc is
always built from source and our changes in meta-iot2000-bsp are no longer required
(reducing the gap with upstream)

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>