meta-iot2000-bsp: add support for user-data
The iot2000 BSP did not include user-data persistent data on a read-only
rootfs as some of our other BSPs do.  Add in support for this feature,
modelled on the MF0200 approach, and store /var/lib (where docker
containers and the package management database are store) to the persistent
data area.

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>